there were two duplicate `create_responder` test methods
